requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
croniter==1.3.7
cryptography==37.0.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-model-utils==4.3.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
fcm-django==1.0.12
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.4
ndg-httpsclient==0.5.1
packaging==21.3
Pillow==9.2.0
proto-plus==1.22.0
protobuf==3.20.1
psycopg2==2.9.3
pusher==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
rq==1.11.0
rq-scheduler==0.11.0
rsa==4.9
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
whitenoise==6.2.0
wrapt==1.14.1
